Use actual migration to test disable_departure!
For some reason, it's impossible to stub `uses_departure` to make it
return `false` anymore, so the integration test for this feature would
be failing on Rails 7.

So, after digging through it for a while, this commit introduces a new
migration that uses `disable_departure!` to disable departure instead of
trying to stub out `uses_departure`.

I think this is a good compromise as this will actually be the way that
users use (and disable) departure in the migration.